requirements.txt
fastapi
uvicorn
groq
langchain
langchain-groq
python-dotenv
pypdf
pytesseract
Pillow
pdf2image
python-multipart
PyMuPDF
langchain
langchain-community
unstructured[all-docs]
pdf2image
langchain-core 
numpy
easyocr
opencv-python
pandas
